requirements.txt
# LAIM - Lab Asset Inventory Manager
# Python Dependencies

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0

# Authentication
passlib==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.18

# Templating
jinja2==3.1.4

# Validation
pydantic==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0

# Utilities
python-dotenv==1.0.1

# HTTP Client (async)
httpx>=0.27.0

# Background Task Scheduler
apscheduler>=3.10.0
